backend/requirements.txt
FastAPI==0.104.1
uvicorn[standard]==0.24.0
SQLAlchemy==2.0.23
Alembic==1.13.1
python-jose[cryptography]==3.3.0
bcrypt==4.1.1
pydantic==2.5.2
pydantic-settings==2.1.0
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
